Moving exported hooks to background-panel Updating props
